requirements.txt
aiohttp
black
coverage
python-dotenv
flake8
mock
pyflakes
pylint
pytest
pytest-asyncio
pytest-cov
python-dotenv
requests
requests-toolbelt
ruff
xmltodict